Implement data-driven layout for Actions
First idea was to add Action schema to the scope, but it lead to
serious performance issues: when any value on the scope changes, all
the widgets which layout depends on it are redrawn, which introduces
form flickering and high latency. Thus schema had to be pushed to the
service, enabling its injection at the directive compile phase (it is
evaluated only once, contrary to the link phase).

Change-Id: I8e2513c5aeabd8ec4bbf72ed025bd3d5da08f94d