Enable changing Task structure
Provide unit-tests for verifying that Task structure depends both on
parent Workflow type and on Task type.

This commits also depends on a change made to Barricade.js library:
now it passes the parameters of a parent container to the children
containers, which add new elements to it via prototype inheritance.

Closes-Bug: #1446228
Implements: blueprint merlin-unittests
Change-Id: Ic4c0539297d6df9a0b1450a824eeca4749455cfd