Added test execution in Firefox.
This adds capturing and executing tests in the Firefox browser. It also
adds the use of the jasmine-jquery framework to permit easier detection
of DOM elements in a rendered directive, as the present method does not
play well with Firefox' tendency to randomly reorder attribute positions.

Testing with Firefox is already supported in infra.

Change-Id: Ifc81e1b243975536ee40e627c1e4a164855fe85c