Add patched version of angular-ui bootstrap lib
This patch adds patched version of angular-ui/bootstrap lib.
We need to patch it because of typeahead - one of bootstrap
components. This component does not support
ngModelOptions="{getterSetter: true}", which we need for
barricade.js integration.
Angular-ui/bootstrap upstream bug:
https://github.com/angular-ui/bootstrap/issues/3823
Until this bug is not fixed in upstream version,
we need to use this patched version of library.

Change-Id: I30a5d52d3a638fa2d7ca110d8824682c3d205611