Use eslint-config-openstack
Instead of explicitly copying all the linting rules between projects,
openstack now publishes a set of rules (much like hacking) that can
be included in projects that want to adhere by our standards. This
patch switches merlin to use this set of rules.

Note: The rules in eslint-config-openstack are a verbatim copy
of what's currently in use in Horizon. Future versions will likely start
activating rules flagged as "TODO", however you can avoid build
instability by avoiding the use of fuzzy version matching in
package.json (which this patch does).

Also, I had to bump down the version of eslint in this patch, because
there was an explicit version conflict. Once we get to the point
of synchronizing dependencies, we'll be able to manage this from a
central location.

Change-Id: I5fa569e314d0b7f936b750e4845295debe5420a8